Geofence_Service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.0
pydantic==2.8.2
shapely==2.0.4
python-dotenv==1.0.1